Adds contentType and Progress enums; Refactoring
